Fix formatting of keywords.txt (#11)
The DATA_TYPE and PREPROCESSOR keyword identifiers are only valid when used in the RSYNTAXTEXTAREA_TOKENTYPE field of keywords.txt. That is the fourth field rather than the second, where they were previously. I moved them to the fourth field and added it with the equivalent KEYWORD_TOKENTYPEs in the second field (only used by Arduino IDE 1.6.4 and older).

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#keywords